structured and updated the jekins file
