+++ .dockerignore
@@ -1,2 +1,5 @@
-.env.sample
-node_modules
+frontend/node_modules
+backend/__pycache__
+.git
+.vscode
+.dockerignore
