+++ .gitignore
@@ -1,2 +1,8 @@
 nginx-ingress
 .env
+
+frontend/node_modules
+backend/__pycache__
+.git
+.vscode
+.dockerignore
