+++ backend/Dockerfile
@@ -1,0 +1,15 @@
+# Stage 1: Python
+FROM python:latest as python-stage
+
+WORKDIR /app
+
+COPY requirements.txt /app/
+
+RUN pip install -r requirements.txt
+
+COPY . /app
+
+EXPOSE 5000
+
+CMD ["python", "main.py"]
+
