+++ backend/requirements.txt
@@ -1,0 +1,4 @@
+Flask
+flask-cors
+requests
+load_dotenv
