+++ frontend/src/App.vue
@@ -1,0 +1,17 @@
+<template>
+  <div id="app">
+        <DomainsTable />
+  </div>
+</template>
+
+<script>
+import DomainsTable from './components/DomainsTable.vue'
+
+export default {
+  name: 'App',
+  components: {
+    DomainsTable
+  }
+};
+</script>
+
