--- ing.yaml
@@ -1,2 +1,0 @@
-controller:
-  replicaCount: 1
