--- nginx.conf
@@ -1,32 +1,0 @@
-server {
-    listen 80;
-    root /usr/share/nginx/html;
-
-    location / {
-        try_files $uri $uri/ /index.html;
-    }
-
-    location = /index.html {
-        expires -1;
-    }
-
-    location /proxy/ {
-        proxy_pass http://localhost:5000/;
-        proxy_set_header Host $host;
-        proxy_set_header X-Real-IP $remote_addr;
-        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-
-        # CORS configuration
-        add_header 'Access-Control-Allow-Origin' '*';
-        add_header 'Access-Control-Request-Private-Network' 'true';
-        add_header 'Access-Control-Allow-Methods' '*';
-        add_header 'Access-Control-Allow-Headers' 'DNT,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Range';
-        add_header 'Access-Control-Allow-Credentials' 'true';
-        add_header 'Access-Control-Max-Age' 1728000;
-        if ($request_method = 'OPTIONS') {
-            add_header 'Content-Type' 'text/plain; charset=utf-8';
-            add_header 'Content-Length' 0;
-            return 204;
-        }
-    }
-}
