--- pythonDockerfile
@@ -1,21 +1,0 @@
-# Stage 1: Python
-FROM python:latest as python-stage
-
-WORKDIR /app
-
-COPY requirements.txt /app/
-
-RUN pip install -r requirements.txt
-
-COPY . /app
-
-ENV ZONE_ID=88212a53b6feba598b197f3508f35b52
-
-ENV CF_API_KEY=ab590d1c5d3139416fef3d173ad4267a75a41
-
-ENV CF_API_EMAIL=safe@hostspaceng.com
-
-ENV VUE_APP_PROXY_URL=http://localhost:5000/
-
-EXPOSE 5000
-CMD ["python", "main.py"]
