--- requirements.txt
@@ -1,4 +1,0 @@
-Flask
-flask-cors
-requests
-load_dotenv
